Implemented email verification, user authorization, login.
Ight there is a lot to say here. PROBLEMS:
- dupe list_events() -> idk which one should stay, I dont remember who made which and dont want to delete anyone's work
- utcnow() -> whoever did one list_events() and next_event_for_club used utcnow() which my vscode is flagging the hell out of.
As for what is going on here I did everything and also added a toggle for a couple things. The FRONTEND_BASE and BACKEND_BASE obviously point towards local rn. We will change that once we deploy.
Also for the verification, I added a "dev" mode so raghav can test whether everything works. This simply returns the url. If you change _EMAIL_MODE into "smtp", it sends the email.